+++ src/backend/marsha/core/forms.py
@@ -1,7 +1,9 @@
 """Marsha forms module."""
-from django.forms import ModelForm
+from django.core.exceptions import ValidationError
+from django.forms import CharField, ModelForm
 
 from . import models
+from .defaults import INITIALIZED
 
 
 class DocumentForm(ModelForm):
@@ -17,8 +19,29 @@ class DocumentForm(ModelForm):
 class VideoForm(ModelForm):
     """Form to create or update videos."""
 
+    upload_state = CharField(
+        max_length=20,
+        required=False,
+    )
+
     class Meta:
         """Meta for VideoForm."""
 
         model = models.Video
-        fields = ["description", "is_public", "lti_id", "playlist", "title"]
+        fields = [
+            "description",
+            "is_public",
+            "lti_id",
+            "playlist",
+            "title",
+            "upload_state",
+        ]
+
+    def clean_upload_state(self):
+        """Check upload_state valid value."""
+        upload_state = self.cleaned_data["upload_state"]
+
+        if upload_state and upload_state != INITIALIZED:
+            raise ValidationError(f"{INITIALIZED} is the only accepted value")
+
+        return upload_state